Cleanup the JS code and packaging process.
