+++ app/public/lib/draw2d_extensions/MVU_messages.js
@@ -4,10 +4,10 @@
  * These functions get initialised by the draw2dWrapper.
  */
 
-let dispatchInferWidthsMessage         = "undefined";
-let dispatchOnSelectComponentMessage   = "undefined";
-let dispatchOnUnselectComponentMessage = "undefined";
-let dispatchHasUnsavedChangesMessage   = "undefined";
+global.dispatchInferWidthsMessage         = "undefined";
+global.dispatchOnSelectComponentMessage   = "undefined";
+global.dispatchOnUnselectComponentMessage = "undefined";
+global.dispatchHasUnsavedChangesMessage   = "undefined";
 
 function setDispatchMessages(
         dispatchInferWidthsMessage_,
@@ -21,10 +21,4 @@ function setDispatchMessages(
     dispatchHasUnsavedChangesMessage   = dispatchHasUnsavedChangesMessage_;
 }
 
-export {
-    setDispatchMessages,
-    dispatchInferWidthsMessage,
-    dispatchOnSelectComponentMessage,
-    dispatchOnUnselectComponentMessage,
-    dispatchHasUnsavedChangesMessage,
-};
+module.exports = { setDispatchMessages };
