+++ package.json
@@ -15,11 +15,9 @@
     "bulma": "^0.7.2",
     "cross-zip": "^3.1.0",
     "css-loader": "^2.1.0",
-    "draw2d": "^1.0.34",
     "electron-packager": "^14.2.1",
     "file-loader": "^3.0.1",
     "font-awesome": "^4.7.0",
-    "jquery": "^3.4.1",
     "node-sass": "^4.12.0",
     "react": "^16.7.0",
     "react-dom": "^16.7.0",
@@ -36,7 +34,6 @@
     "fable-loader": "^1.1.4",
     "fable-utils": "^1.0.6",
     "loglevel": "^1.5.1",
-    "mini-css-extract-plugin": "^0.9.0",
     "uglifyjs-webpack-plugin": "^2.2.0",
     "webpack": "^4.43.0",
     "webpack-cli": "^3.3.11"
