+++ webpack.config.js
@@ -45,9 +45,6 @@ var basicConfig = {
       {
           test: /\.(sass|scss|css)$/,
           use: [
-              //isProduction
-              //    ? MiniCssExtractPlugin.loader
-              //    : 'style-loader',
               'style-loader',
               'css-loader',
               'sass-loader',
@@ -65,12 +62,12 @@ var basicConfig = {
   }
 };
 
-//const UglifyJSPlugin = require('uglifyjs-webpack-plugin');
+// At the moment draw2d and jquery are NOT npm/yarn dependencies. The source
+// code is copied in the app/public/lib/ folder. This should be fine but if you
+// want to update them via the package manager, you need the copy-webpack-plugin
+// to copy the source files to the app/public/lib/ folder while packaging.
 
 var mainConfig = Object.assign({
-  //plugins: [
-  //  new UglifyJSPlugin()
-  //],
   target: "electron-main",
   entry: resolve("src/Main/Main.fsproj"),
   output: {
@@ -80,9 +77,6 @@ var mainConfig = Object.assign({
 }, basicConfig);
 
 var rendererConfig = Object.assign({
-  //plugins: [
-  //  new UglifyJSPlugin()
-  //],
   target: "electron-renderer",
   entry: [ resolve("src/Renderer/Renderer.fsproj") ],
   output: {
